move plugins to non-tested folder, remove hardcoded integrations
