--- config/plugins.go
@@ -1,16 +1,0 @@
-package config
-
-type Plugins struct {
-	Twilio    Twilio
-	PagerDuty PagerDuty
-}
-
-type Twilio struct {
-	AccountID string
-	Auth      string
-}
-
-type PagerDuty struct {
-	Email string
-	Key   string
-}
